refactor(core): keep one codec for the run summary
The `<<run-summary>>` payload was serialized and re-parsed in three places,
each with its own guards, right after the format moved from markdown to
JSON. Serialization and structured parsing now live next to the RunSummary
type in core, and both the store writer and the recap component consume
them. The legacy markdown fallback hangs off the shared parser returning
null, so old rows still render without a migration.
