+++ packages/core/src/orchestrator/parser.ts
@@ -2,6 +2,7 @@ import type { ModelEffort, ProviderId } from '@goodboy/types';
 import { providerEffortLevels } from '../providers/providerEffortLevels';
 import { resolveStoredModelSelection } from '../providers/resolveStoredModelSelection';
 import { isAgentRole } from '../roles';
+import { structuredRunSummary } from './runSummary';
 import type { OrchestratorDecision, OrchestratorStep, RunSummary } from './types';
 
 const START_MARKER = '<<orchestrator>>';
@@ -95,25 +96,12 @@ const runSummarySlice = (raw: string): string | null => {
   return nonEmptyString(raw.slice(contentStart, end));
 };
 
-const stringList = (value: unknown): ReadonlyArray<string> =>
-  Array.isArray(value)
-    ? value.map((entry) => nonEmptyString(entry)).filter((entry): entry is string => entry !== null)
-    : [];
-
 const parseRunSummary = (raw: string): RunSummary | null => {
   const slice = runSummarySlice(raw);
   if (slice === null) {
     return null;
   }
-  const parsed = asRecord(parseJson(slice));
-  if (parsed === null || (!Array.isArray(parsed['done']) && !Array.isArray(parsed['left']))) {
-    return { kind: 'text', text: slice };
-  }
-  return {
-    kind: 'structured',
-    done: stringList(parsed['done']),
-    left: stringList(parsed['left']),
-  };
+  return structuredRunSummary(parseJson(slice)) ?? { kind: 'text', text: slice };
 };
 
 type ModelParams = {
